fix(server): reject auth headers on nango_proxy and undo the file merge
Caller headers are merged over the credentials the proxy builds from the
connection, and only user-agent is protected, so an agent could set
authorization and have the request go out as something other than the
session's connection. The tool now rejects those and the hop-by-hop
headers instead of forwarding them, which is what its description
already claimed. Only the session tool rejects them, proxy_request is
reached with an API key that already holds the environment.

integration takes providerConfigKeySchema now, same as proxy_request,
so the tool stops advertising keys that cannot resolve.

Splits mcpProxy.service.ts back up. Moving the shared execution out of
the management controller was the point, folding formatter and response
into one file with it was not, and it buried 240 lines of unchanged code
plus a reworded limit message in a diff that read as new.
