+++ packages/server/lib/controllers/mcp/proxy/request.unit.test.ts
@@ -4,7 +4,7 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { Err, Ok } from '@nangohq/utils';
 
-import { MAX_MCP_PROXY_RESPONSE_BYTES } from '../../../services/mcpProxy.service.js';
+import { MAX_MCP_PROXY_RESPONSE_BYTES } from '../../../services/mcpProxyResponse.js';
 import proxyService, { ProxyServiceError } from '../../../services/proxy.service.js';
 import { egressTelemetryRecorder } from '../../../utils/egressTelemetry.js';
 import { PublicMcpError } from '../utils.js';
